// src/services/openApi/models/Smile.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Smile = {
    advantages?: string;
    callDate?: string;
    choiceReason?: string;
    complete?: string;
    createdAt?: string;
    disadvantages?: string;
    endDate?: string;
    giveName?: string;
    id?: number;
    phone?: string;
    reaction?: number;
    resignationReason?: string;
    resignationReasonCoincide?: string;
    returnPossibility?: string;
    significant?: string;
    takeName?: string;
    workNow?: string;
    writer?: string;
};